[Fix] 모바일 프로그램 신청 이어서 신청하기/내용 유지 적용
