+++ src/components/common/program/program-detail/apply/content/InputContent.tsx
@@ -88,10 +88,10 @@ const InputContent = ({
   }, [scrollableBoxRef]);
 
   return (
-    <div className="flex max-h-[36rem] flex-col gap-5 md:h-full">
+    <div className="flex flex-col gap-5">
       <ScrollableBox
         ref={scrollableBoxRef}
-        className={clsx('overflow-y flex h-full flex-col gap-3', {
+        className={clsx('flex h-full flex-col gap-3', {
           'max-h-[30rem]': mobileMatches,
           'max-h-[40rem]': tabletMatches,
         })}
