+++ src/components/common/program/program-detail/apply/content/PayContent.tsx
@@ -1,11 +1,11 @@
 import { useEffect, useRef } from 'react';
 import { FaArrowLeft } from 'react-icons/fa6';
 
+import { PayInfo } from '../../section/ApplySection';
+import ScrollableBox from '../scrollable-box/ScrollableBox';
 import CouponSection from '../section/CouponSection';
 import PayInfoSection from '../section/PayInfoSection';
 import PriceSection from '../section/PriceSection';
-import ScrollableBox from '../scrollable-box/ScrollableBox';
-import { PayInfo } from '../../section/ApplySection';
 
 interface ScrollableDiv extends HTMLDivElement {
   scrollTimeout?: number;
